fix(types): optimize export of typings
