+++ packages/visual-sdk/src/index.ts
@@ -1,1 +1,1 @@
-export * from './store';
+export * from './store/index';
